vulkan-hdr-layer-kwin6: init at 0-unstable-2024-10-19
https://github.com/Zamundaaa/VK_hdr_layer

Enables HDR support for Vulkan clients (such as mpv) on kwin6.

This is likely to be a temporary solution until Wayland upstream
adopts the necessary protocols, however that has been in the works for
many years:

https://gitlab.freedesktop.org/wayland/wayland-protocols/-/merge_requests/14

For the time being, this provides a path for users to experiment with
HDR, share feedback and hopefully make progress.

Arch provides a similar package:
https://aur.archlinux.org/packages/vk-hdr-layer-kwin6-git